tests/wpt/tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_popups_nonescaping-2.html

<!doctype html>
<meta charset=utf-8>
<title>Check that popups from a sandboxed iframe do not escape the sandbox</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
  var t = async_test();
  onmessage = t.step_func_done(function(e) {
    assert_equals(e.origin, "null", "It came from a sandboxed iframe");
    assert_equals(e.data.data, undefined, "Should have the right message");
    assert_equals(e.data.origin, "null", "Should not have escaped the sandbox");
  });
  addEventListener("load", function() {
    frames[0].postMessage("start", "*");
  });
</script>
<iframe sandbox="allow-scripts allow-popups"
        src="iframe_sandbox_popups_helper-2.html"></iframe>
